+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+v1.2
+----
+- Add of Timezone config as microtime use UTC (17/08/2017)
+- Creation of a `reUse()` method to create payment form for a payment initiated but not executed
+
 v1.1
 ----
 - Remove of .travis.yml as tests have to be defined before (01/08/2017)